Move some RichText e2e tests
