+++ packages/e2e-tests/specs/rtl.test.js
@@ -17,24 +17,6 @@ describe( 'RTL', () => {
 		await createNewPost();
 	} );
 
-	it( 'should arrow navigate', async () => {
-		await page.evaluate( () => document.dir = 'rtl' );
-		await page.keyboard.press( 'Enter' );
-
-		// We need at least three characters as arrow navigation *from* the
-		// edges might be handled differently.
-		await page.keyboard.type( ARABIC_ONE );
-		await page.keyboard.type( ARABIC_TWO );
-		await page.keyboard.press( 'ArrowRight' );
-		// This is the important key press: arrow nav *from* the middle.
-		await page.keyboard.press( 'ArrowRight' );
-		await page.keyboard.type( ARABIC_ZERO );
-
-		// Expect: ARABIC_ZERO + ARABIC_ONE + ARABIC_TWO (<p>٠١٢</p>).
-		// N.b.: HTML is LTR, so direction will be reversed!
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
 	it( 'should split', async () => {
 		await page.evaluate( () => document.dir = 'rtl' );
 		await page.keyboard.press( 'Enter' );
@@ -98,25 +80,5 @@ describe( 'RTL', () => {
 		await pressKeyWithModifier( 'shift', 'Enter' );
 
 		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'should navigate inline boundaries', async () => {
-		await page.evaluate( () => document.dir = 'rtl' );
-		await page.keyboard.press( 'Enter' );
-
-		await pressKeyWithModifier( 'primary', 'b' );
-		await page.keyboard.type( ARABIC_ONE );
-		await pressKeyWithModifier( 'primary', 'b' );
-		await page.keyboard.type( ARABIC_TWO );
-
-		// Insert a character at each boundary position.
-		for ( let i = 4; i > 0; i-- ) {
-			await page.keyboard.press( 'ArrowRight' );
-			await page.keyboard.type( ARABIC_ZERO );
-
-			expect( await getEditedPostContent() ).toMatchSnapshot();
-
-			await page.keyboard.press( 'Backspace' );
-		}
 	} );
 } );
